pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pred-maint-arena"
version = "1.0.0"
description = "OpenEnv: Predictive Maintenance Arena — AI agents manage factory machines"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "websockets>=12.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
training = ["trl>=0.12.0", "transformers>=4.40.0", "torch>=2.0.0"]
dev = ["pytest>=7.0", "httpx>=0.25.0"]

[tool.setuptools.packages.find]
include = ["pred_maint_arena*", "server*"]
